Issue #40: Update DB with new version for new photo selection method.
